requirements.txt
discord
ruamel.yaml
pydantic
asyncpg